Refactored after tech spec change.
